feat: fix thundering herd problem in rate limiter creation
- Add DashMap for per-configuration creation locks
- Implement double-checked locking pattern to ensure only one thread creates a rate limiter
- Clean up locks after use to prevent memory leaks
- Fix string interpolation in log messages per project guidelines
- Move rate limiting tests to separate module

This prevents multiple threads from creating duplicate rate limiters for the
same configuration when they all miss the cache simultaneously.
