+++ crates/rate-limit/Cargo.toml
@@ -7,6 +7,7 @@ license.workspace = true
 [dependencies]
 anyhow.workspace = true
 config.workspace = true
+dashmap.workspace = true
 deadpool.workspace = true
 governor.workspace = true
 log.workspace = true
