+++ crates/rate-limit/src/request.rs
@@ -34,32 +34,32 @@ impl RateLimitRequestBuilder {
         self.ip = Some(ip);
         self
     }
-    
+
     /// Set the IP address from a string.
     pub fn ip_str(mut self, ip: &str) -> Result<Self, std::net::AddrParseError> {
         self.ip = Some(ip.parse()?);
         Ok(self)
     }
-    
+
     /// Set the server name.
     pub fn server(mut self, name: impl Into<String>) -> Self {
         self.server_name = Some(name.into());
         self
     }
-    
+
     /// Set the tool name.
     pub fn tool(mut self, name: impl Into<String>) -> Self {
         self.tool_name = Some(name.into());
         self
     }
-    
+
     /// Set both server and tool name at once.
     pub fn server_tool(mut self, server: impl Into<String>, tool: impl Into<String>) -> Self {
         self.server_name = Some(server.into());
         self.tool_name = Some(tool.into());
         self
     }
-    
+
     /// Build the rate limit request.
     pub fn build(self) -> RateLimitRequest {
         RateLimitRequest {
@@ -69,4 +69,3 @@ impl RateLimitRequestBuilder {
         }
     }
 }
-
