+++ crates/rate-limit/src/storage/redis_pool.rs
@@ -3,7 +3,7 @@
 use std::sync::atomic::{AtomicUsize, Ordering};
 
 use deadpool::managed::{self, Metrics};
-use redis::{aio::MultiplexedConnection, Client, RedisError, RedisResult};
+use redis::{Client, RedisError, RedisResult, aio::MultiplexedConnection};
 
 use config::{RedisConfig, RedisTlsConfig};
 
@@ -23,11 +23,8 @@ impl Manager {
         let client = if let Some(tls_config) = &config.tls {
             // For Redis with TLS, build certificates or use insecure mode
             let tls_certs = build_tls_certificates(tls_config)?;
-            
-            Client::build_with_tls(
-                config.url.clone(),
-                tls_certs,
-            )?
+
+            Client::build_with_tls(config.url.clone(), tls_certs)?
         } else {
             Client::open(config.url.as_str())?
         };
@@ -48,11 +45,7 @@ impl managed::Manager for Manager {
         Ok(conn)
     }
 
-    async fn recycle(
-        &self,
-        conn: &mut MultiplexedConnection,
-        _: &Metrics,
-    ) -> managed::RecycleResult<Self::Error> {
+    async fn recycle(&self, conn: &mut MultiplexedConnection, _: &Metrics) -> managed::RecycleResult<Self::Error> {
         let ping_number = self.ping_number.fetch_add(1, Ordering::Relaxed).to_string();
 
         let (n,) = redis::Pipeline::with_capacity(2)
@@ -74,31 +67,38 @@ impl managed::Manager for Manager {
 /// Build TLS certificates from configuration.
 fn build_tls_certificates(config: &RedisTlsConfig) -> RedisResult<redis::TlsCertificates> {
     use redis::ClientTlsConfig;
-    
+
     // For insecure mode, we'll use the CA cert from the container
     if config.insecure.unwrap_or(false) {
         // Try to load the CA cert for self-signed certificates
-        let ca_path = config.ca_cert_path.as_deref()
+        let ca_path = config
+            .ca_cert_path
+            .as_deref()
             .unwrap_or("./crates/integration-tests/docker/redis/tls/ca.crt");
-            
+
         let root_cert = std::fs::read(ca_path).ok();
-        
+
         return Ok(redis::TlsCertificates {
             client_tls: None,
             root_cert,
         });
     }
-    
+
     let mut client_tls = None;
     let mut root_cert = None;
 
     // Load client certificate and key if provided
     if let (Some(cert_path), Some(key_path)) = (&config.client_cert_path, &config.client_key_path) {
-        let cert = std::fs::read(cert_path)
-            .map_err(|e| RedisError::from((redis::ErrorKind::IoError, "Failed to read client certificate", e.to_string())))?;
+        let cert = std::fs::read(cert_path).map_err(|e| {
+            RedisError::from((
+                redis::ErrorKind::IoError,
+                "Failed to read client certificate",
+                e.to_string(),
+            ))
+        })?;
         let key = std::fs::read(key_path)
             .map_err(|e| RedisError::from((redis::ErrorKind::IoError, "Failed to read client key", e.to_string())))?;
-        
+
         client_tls = Some(ClientTlsConfig {
             client_cert: cert,
             client_key: key,
@@ -107,43 +107,45 @@ fn build_tls_certificates(config: &RedisTlsConfig) -> RedisResult<redis::TlsCert
 
     // Load CA certificate if provided
     if let Some(ca_path) = &config.ca_cert_path {
-        root_cert = Some(std::fs::read(ca_path)
-            .map_err(|e| RedisError::from((redis::ErrorKind::IoError, "Failed to read CA certificate", e.to_string())))?);
+        root_cert = Some(std::fs::read(ca_path).map_err(|e| {
+            RedisError::from((
+                redis::ErrorKind::IoError,
+                "Failed to read CA certificate",
+                e.to_string(),
+            ))
+        })?);
     }
 
-    Ok(redis::TlsCertificates {
-        client_tls,
-        root_cert,
-    })
+    Ok(redis::TlsCertificates { client_tls, root_cert })
 }
 
 /// Create a Redis connection pool from configuration.
 pub fn create_pool(config: &RedisConfig) -> RedisResult<Pool> {
     let manager = Manager::new(config)?;
-    
+
     let mut pool_config = deadpool::managed::PoolConfig::default();
-    
+
     if let Some(max_size) = config.pool.max_size {
         pool_config.max_size = max_size;
     }
-    
+
     if let Some(timeout_create) = config.pool.timeout_create {
         pool_config.timeouts.create = Some(timeout_create);
     }
-    
+
     if let Some(timeout_wait) = config.pool.timeout_wait {
         pool_config.timeouts.wait = Some(timeout_wait);
     }
-    
+
     if let Some(timeout_recycle) = config.pool.timeout_recycle {
         pool_config.timeouts.recycle = Some(timeout_recycle);
     }
-    
+
     let pool = Pool::builder(manager)
         .config(pool_config)
         .runtime(deadpool::Runtime::Tokio1)
         .build()
         .map_err(|e| RedisError::from((redis::ErrorKind::IoError, "Failed to create pool", e.to_string())))?;
-    
+
     Ok(pool)
 }
